Refactor host and port parsing
